Clean old/new files
