+++ bandits/base_model.py
