+++ bandits/nig_normal.py
