+++ bandits/output_format.py
