--- pg_gan/metrics/__init__.py
@@ -1,1 +1,0 @@
-# empty
