--- pg_gan/metrics/inception_score.py
@@ -1,147 +1,0 @@
-# Copyright 2016 Wojciech Zaremba
-#
-# Licensed under the Apache License, Version 2.0 (the "License");
-# you may not use this file except in compliance with the License.
-# You may obtain a copy of the License at
-#
-#     http://www.apache.org/licenses/LICENSE-2.0
-#
-# Unless required by applicable law or agreed to in writing, software
-# distributed under the License is distributed on an "AS IS" BASIS,
-# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-# See the License for the specific language governing permissions and
-# limitations under the License.
-
-# Adapted from the original implementation by Wojciech Zaremba.
-# Source: https://github.com/openai/improved-gan/blob/master/inception_score/model.py
-
-from __future__ import absolute_import
-from __future__ import division
-from __future__ import print_function
-
-import os.path
-import sys
-import tarfile
-
-import numpy as np
-from six.moves import urllib
-import tensorflow as tf
-import glob
-import scipy.misc
-import math
-import sys
-
-MODEL_DIR = '/tmp/imagenet'
-
-DATA_URL = 'http://download.tensorflow.org/models/image/imagenet/inception-2015-12-05.tgz'
-softmax = None
-
-# Call this function with list of images. Each of elements should be a 
-# numpy array with values ranging from 0 to 255.
-def get_inception_score(images, splits=10):
-  assert(type(images) == list)
-  assert(type(images[0]) == np.ndarray)
-  assert(len(images[0].shape) == 3)
-  #assert(np.max(images[0]) > 10) # EDIT: commented out
-  #assert(np.min(images[0]) >= 0.0)
-  inps = []
-  for img in images:
-    img = img.astype(np.float32)
-    inps.append(np.expand_dims(img, 0))
-  bs = 100
-  with tf.Session() as sess:
-    preds = []
-    n_batches = int(math.ceil(float(len(inps)) / float(bs)))
-    for i in range(n_batches):
-        #sys.stdout.write(".") # EDIT: commented out
-        #sys.stdout.flush()
-        inp = inps[(i * bs):min((i + 1) * bs, len(inps))]
-        inp = np.concatenate(inp, 0)
-        pred = sess.run(softmax, {'ExpandDims:0': inp})
-        preds.append(pred)
-    preds = np.concatenate(preds, 0)
-    scores = []
-    for i in range(splits):
-      part = preds[(i * preds.shape[0] // splits):((i + 1) * preds.shape[0] // splits), :]
-      kl = part * (np.log(part) - np.log(np.expand_dims(np.mean(part, 0), 0)))
-      kl = np.mean(np.sum(kl, 1))
-      scores.append(np.exp(kl))
-    return np.mean(scores), np.std(scores)
-
-# This function is called automatically.
-def _init_inception():
-  global softmax
-  if not os.path.exists(MODEL_DIR):
-    os.makedirs(MODEL_DIR)
-  filename = DATA_URL.split('/')[-1]
-  filepath = os.path.join(MODEL_DIR, filename)
-  if not os.path.exists(filepath):
-    def _progress(count, block_size, total_size):
-      sys.stdout.write('\r>> Downloading %s %.1f%%' % (
-          filename, float(count * block_size) / float(total_size) * 100.0))
-      sys.stdout.flush()
-    filepath, _ = urllib.request.urlretrieve(DATA_URL, filepath, _progress)
-    print()
-    statinfo = os.stat(filepath)
-    print('Succesfully downloaded', filename, statinfo.st_size, 'bytes.')
-    tarfile.open(filepath, 'r:gz').extractall(MODEL_DIR) # EDIT: increased indent
-  with tf.gfile.FastGFile(os.path.join(
-      MODEL_DIR, 'classify_image_graph_def.pb'), 'rb') as f:
-    graph_def = tf.GraphDef()
-    graph_def.ParseFromString(f.read())
-    _ = tf.import_graph_def(graph_def, name='')
-  # Works with an arbitrary minibatch size.
-  with tf.Session() as sess:
-    pool3 = sess.graph.get_tensor_by_name('pool_3:0')
-    ops = pool3.graph.get_operations()
-    for op_idx, op in enumerate(ops):
-        for o in op.outputs:
-            shape = o.get_shape()
-            shape = [s.value for s in shape]
-            new_shape = []
-            for j, s in enumerate(shape):
-                if s == 1 and j == 0:
-                    new_shape.append(None)
-                else:
-                    new_shape.append(s)
-            try:
-                o._shape = tf.TensorShape(new_shape)
-            except ValueError:
-                o._shape_val = tf.TensorShape(new_shape) # EDIT: added for compatibility with tensorflow 1.6.0
-    w = sess.graph.get_operation_by_name("softmax/logits/MatMul").inputs[1]
-    logits = tf.matmul(tf.squeeze(pool3), w)
-    softmax = tf.nn.softmax(logits)
-
-#if softmax is None: # EDIT: commented out
-#  _init_inception() # EDIT: commented out
-
-#----------------------------------------------------------------------------
-# EDIT: added
-
-class API:
-    def __init__(self, num_images, image_shape, image_dtype, minibatch_size):
-        import config
-        globals()['MODEL_DIR'] = os.path.join(config.result_dir, '_inception')
-        self.sess = tf.get_default_session()
-        _init_inception()
-
-    def get_metric_names(self):
-        return ['IS_mean', 'IS_std']
-
-    def get_metric_formatting(self):
-        return ['%-10.4f', '%-10.4f']
-
-    def begin(self, mode):
-        assert mode in ['warmup', 'reals', 'fakes']
-        self.images = []
-
-    def feed(self, mode, minibatch):
-        self.images.append(minibatch.transpose(0, 2, 3, 1))
-
-    def end(self, mode):
-        images = list(np.concatenate(self.images))
-        with self.sess.as_default():
-            mean, std = get_inception_score(images)
-        return [mean, std]
-
-#----------------------------------------------------------------------------
