--- templates/webUI.html
@@ -1,100 +1,0 @@
-<html>
-<head>
-    <title>JAMM Bandit</title>
-<meta name="viewport" content="width=device-width, height=device-height, initial-scale=1.0, maximum-scale=1.0, minimum-scale=1.0">
-<link rel="stylesheet" href="https://maxcdn.bootstrapcdn.com/bootstrap/4.0.0/css/bootstrap.min.css" integrity="sha384-Gn5384xqQ1aoWXA+058RXPxPg6fy4IWvTNh0E263XmFcJlSAwiGgFAW/dAiS6JXm" crossorigin="anonymous">
-
-</head>
-
-<script>
-
-function noChange() {
-    var imgElement = document.getElementById('candidate');
-	imgElement.src = "../images/2.png";
-}
-
-function yesChange() {
-    var imgElement = document.getElementById('candidate');
-	imgElement.src = "../images/3.png";
-}
-
-function skip() {
-    var imgElement = document.getElementById('candidate');
-	imgElement.src = "../images/4.png";
-}
-
-/*
-// user_choice is an arbitrary string label or int. 
-function result(user_choice) {
-  console.log(user_choice);
-  image = document.getElementById("candidate");
-if!image) { 
-    console.log("No image element in page."); 
-    return; 
-  }
-  // Load a new image but report result.
-  document.location = "/?log="+image.src+":"+user_choice;
-}
-
-// Simple click
-// document.addEventListener('click', function(){result(0);}, false);        
-
-// Click and drag (non-touch-screen)
-document.addEventListener('dragstart', handleDragStart, false);
-document.addEventListener('dragend', handleDragEnd, false);
-
-// On mobile swipe action
-document.addEventListener('touchstart', handleTouchStart, false);        
-document.addEventListener('touchmove', handleTouchEnd, false);
-
-var xDown = null;                                                        
-function handleDragStart(evt) {                                         
-  xDown = evt.clientX;                                      
-};                                                
-function handleTouchStart(evt) {                                         
-  xDown = evt.touches[0].clientX;                                      
-};                                                
-function handleTouchEnd(evt) {
-  handleEnd(evt.touches[0].clientX);
-}
-function handleDragEnd(evt) {
-  handleEnd(evt.clientX);
-}
-function handleEnd(xUp) {
-  if ( !xDown ) { return; }
-  var xDiff = xDown - xUp;
-  if ( xDiff > 2 ) result(1);   
-  if ( xDiff < -2 ) result(-1);   
-  xDown = null;
-};
-*/
-
-</script>
-
-
-<body class="bg-light" dragable="true" height=100%> 
-
-
-
-  <div class="page-header row justify-content-center">
-    <h1>Is this face <b><i>warm</i></b>?</h1>
-  </div>
-
-  <div class="row justify-content-center">
-	  <img src="{{  image  }}" id="candidate" style="500px", height="500px"> 
-  </div>
-
-  <div class="row justify-content-center">
-    <button type="button" class="btn btn-danger" onclick="noChange();">No</button>
-    <button type="button" class="btn btn-success" onclick="yesChange();">Yes</button>
-  </div>
-  
-
-	
-	<div class="row justify-content-center">
-		<button type="button" class="btn btn-dark" onclick="skip();">There's something wrong with this image (skip)</button>
-	</div>
-
-	
-</body>
-</html>
